scripts/requirements.txt
fastapi
firebase
newspaper3k
jsonlib
bert-summarizer
beautifulsoup4
requests
tweepy
config
selenium
webdriver-manager
uvicorn